% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMcorrProbit.R
\name{emcorrprob}
\alias{emcorrprob}
\title{Estimates via EM.}
\usage{
emcorrprob(xfixed, xrand, ...)
}
\arguments{
\item{xfixed}{Fixed predictors.}

\item{xrand}{Random predictors.}

\item{y}{Ordinal data.}
}
\value{
The estimation of correlated probit model.
}
\description{
Estimates via EM.
}
\examples{
EMcorrProbit(xfixed, xrand, y, start.values.beta, start.values.delta=NULL,  start.values.sigma.rand, exact, montecarlo=100, epsilon=.001, ...)
}

