\name{EMcorrProbit-internal}
\alias{emcorrprobitFit}
\alias{emcorrprobitFit.default}
\alias{emcorrprobitFit.oneord}
\alias{ecm.one.ordinal}
\alias{standard.error.bootstrap.one.ordinal}
\title{Internal functions in EMcorrProbit package}
\description{These are nternal functions in EMcorrProbit package}
