% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMcorrProbit.R
\name{emcorrprobit}
\alias{emcorrprobit}
\title{Estimates via EM.}
\usage{
emcorrprobit(y, ...)
}
\arguments{
\item{y}{Ordinal data.}

\item{xfixed}{Fixed predictors.}

\item{xrand}{Random predictors.}
}
\value{
The estimation of correlated probit model.
}
\description{
Estimates via EM.
}
\examples{
EMcorrProbit(y, xfixed, xrand, start.values.beta, start.values.delta=NULL,  start.values.sigma.rand, exact, montecarlo=100, epsilon=.001, ...)
}

